PolyKybd: 32-bit glyph codepoints (drop emoji PUA shift)
Use real 32-bit Unicode codepoints throughout the per-keycap glyph
pipeline so SMP emoji (> 0xFFFF) no longer have to be shifted into the
BMP Private Use Area.

- GFXfont.first/last widened to uint32_t (base/fonts/gfxfont.h)
- kdisp_write_gfx_char/_text now take char32 (const uint32_t*);
  first/last read via pgm_read_dword
- display strings are char32 (U"..."): named_glyphs.h, lang_lut.c and
  uni.h cog templates emit the real codepoint (xlsx column B) as U"\x..."
- emoji_layer drops the cp_to_font_idx() PUA shift (passes the real
  codepoint straight through)
- oled_helper num/hex_to_u16_string -> _u32_string (char32 output)
- create_fonts.sh: emoji fonts generated with -b32 (real codepoints);
  display-only UI icon fonts keep their PUA addressing; Emoji1 demoted to
  lowest precedence so its multi-range gaps don't shadow the now
  real-codepoint category fonts
- regenerated all SMP emoji font headers and cog tables

split72 and corne42 build clean; every typed emoji, UI icon and BMP
glyph remains covered by a font.

+++ keyboards/handwired/polykybd/base/text_helper.h
@@ -2,4 +2,4 @@
 
 #include QMK_KEYBOARD_H
 
-const uint16_t* get_led_matrix_text(uint8_t rgb_mode);
+const uint32_t* get_led_matrix_text(uint8_t rgb_mode);
